PA

Description

Main module for interface simulation of ATS and PA

Documentation

serverATS

Arguments

:: ServiceName
Port number

-> Scenario
Scenario

-> IO ()

Start simulator for ATS server with scenario Currently following scenarios are available scenario1, scenario2, scenario3, scenario4 and scenario5

serverPA :: HostName -> ServiceName -> IO ()

Start simulator for PA server

getHandle

Arguments

:: HostName
IP Address of the target server to communicate

-> ServiceName
Port number of the target server to communicate

-> IO Handle

Get a handle for TCP/IP communicatoin for manually sending the message with sendMsgPA2ATS or sendMsgATS2PA

closeHandle :: Handle -> IO ()

To close handle

sendMsgPA2ATS

Arguments

:: Handle
Handle of the TCP/IP communication

-> MessagePA2ATS
Message to be sent

-> IO ()

To manually send the PA to ATS message

sendMsgATS2PA

Arguments

:: Handle
Handle of the TCP/IP communication

-> MessagePA2ATS
Message to be sent

-> IO ()

To manually send the ATS to PA message

wait

Arguments

:: Int
seconds

-> IO ()

To insert delay for the thread